requirement.txt
djangorestframework-simplejwt==4.6.0
django
djangorestframework==3.11.0
